Fix class game

+++ game_author.rb
@@ -2,7 +2,6 @@ require 'json'
 require_relative 'game'
 require_relative 'author'
 require_relative 'app'
-require 'date'
 
 class GameStore
   attr_reader :games, :authors
@@ -10,13 +9,12 @@ class GameStore
   def initialize
     @games = []
     @authors = []
-    load_data
   end
 
   def add_game(game)
     @games << game
     game.authors.each { |author| add_author(author) }
-    puts "Game '#{game.title}' has been added ✅"
+    puts "Game '#{game.title}' has been added."
     save_data
   end
 
@@ -28,7 +26,7 @@ class GameStore
 
   def list_games
     if games.empty?
-      puts 'No games found in the catalog! 🙁'
+      puts 'There are no games in the catalog.'
     else
       puts "***************** Games Information 🎮 ********************\n"
       games.each do |game|
@@ -45,7 +43,7 @@ class GameStore
 
   def list_authors
     if authors.empty?
-      puts 'There are no authors in the catalog.'
+      puts 'There are no authors in the catalog.
     else
       puts "***************** Author Information 🧑‍🏫 ********************\n"
       authors.each do |author|
@@ -58,6 +56,8 @@ class GameStore
   end
 
   def save_data
+    return unless File.exist?('./data/games.json') && File.exist?('./data/authors.json')
+
     File.write('./data/games.json', JSON.generate(games.map(&:to_hash)))
     File.write('./data/authors.json', JSON.generate(authors.map(&:to_hash)))
   end
@@ -67,57 +67,61 @@ class GameStore
 
     games_data = JSON.parse(File.read('./data/games.json'), object_class: Game)
     authors_data = JSON.parse(File.read('./data/authors.json'), object_class: Author)
-    @games = games_data.map do |game_data|
-      authors = game_data['authors'].map { |author_data| Author.new(author_data) }
-      Game.new(game_data['title'], game_data['multiplayer'], Date.parse(game_data['last_played_at']), Date.parse(game_data['publish_date']), authors)
-    end
-    authors_data.each do |author_data|
-      author = Author.new(author_data)
-      author.items = games.select { |game| game.authors.include?(author) }
-      add_author(author)
-    end
-  end
-end
 
-def run
-  app = App.new
-  loop do
-    puts "Welcome! 👋😃\nChoose an option: 👉"
-    puts '1. List all games'
-    puts '2. List all authors'
-    puts '3. Add game'
-    puts '4. Quit'
-    choice = gets.chomp.to_i
-    case choice
-    when 1
-      list_games
-    when 2
-      list_authors
-    when 3
-      puts 'Enter game title:'
-      title = gets.chomp
-      puts 'Is the game multiplayer? (Y/N)'
-      multiplayer = gets.chomp.downcase == 'y'
-      puts 'Enter the date of the last time the
+
+    @games = games_data.map do |game_data|
+      game = Game.new(game_data['title'], game_data['multiplayer'], game_data['last_played_at'], game_data['publish_date'])
+      game_data['authors'].each do |author_data|
+        author = Author.new(author_data['first_name'], author_data['last_name'])
+        game.add_author(author)
+      end
+      game
+    end
+
+
+    @authors = authors_data
+  end
+
+  def run
+    app = App.new
+    loop do
+      puts "Welcome! 👋😃\nChoose an option: 👉"
+      puts '1. List all games'
+      puts '2. List all authors'
+      puts '3. Add game'
+      puts '4. Quit'
+      choice = gets.chomp.to_i
+      case choice
+      when 1
+        list_games
+      when 2
+        list_authors
+      when 3
+        puts 'Enter game title:'
+        title = gets.chomp
+        puts 'Is the game multiplayer? (Y/N)'
+        multiplayer = gets.chomp.downcase == 'y'
+        puts 'Enter the date of the last time the
    game was played (YYYY/MM/DD):'
-      last_played_at = gets.chomp
-      puts 'Enter the game\'s publish date (YYYY/MM/DD):'
-      publish_date = gets.chomp
-      game = Game.new(title, multiplayer, last_played_at, publish_date, [])
-      puts 'Enter author first name:'
-      first_name = gets.chomp
-      puts 'Enter author last name:'
-      last_name = gets.chomp
-      author = Author.new(first_name, last_name)
-      game.add_author(author)
-      add_game(game)
-    when 4
-      print 'Thanks for using our catalog 😘'
-      puts "\n"
-      return app.main_menu
-    else
-      puts 'Ooops!!! Invalid option ❌'
-      return run
+        last_played_at = gets.chomp
+        puts 'Enter the game\'s publish date (YYYY/MM/DD):'
+        publish_date = gets.chomp
+        game = Game.new(title, multiplayer, last_played_at, publish_date, [])
+        puts 'Enter author first name:'
+        first_name = gets.chomp
+        puts 'Enter author last name:'
+        last_name = gets.chomp
+        author = Author.new(first_name, last_name)
+        game.add_author(author)
+        add_game(game)
+      when 4
+        print 'Thanks for using our catalog 😘'
+        puts "\n"
+        return app.main_menu
+      else
+        puts 'Ooops!!! Invalid option ❌'
+        return run
+      end
     end
   end
 end
